[Frontend] Add ShelfBookAddComponent

+++ frontend/bookmind/src/app/components/shelf/shelf-details/shelf-details.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from "@angular/router";
 import { defaultShelf } from "../../../domain/default/defaultShelf";
 import { ShelfService } from "../../../service/shelf.service";
+import { ConfirmService } from "../../../service/confirm.service";
 
 @Component({
     selector: 'app-shelf-details',
@@ -9,9 +10,11 @@ import { ShelfService } from "../../../service/shelf.service";
     styleUrls: ['./shelf-details.component.scss']
 })
 export class ShelfDetailsComponent implements OnInit {
+
     public shelf = defaultShelf();
 
     constructor(private act: ActivatedRoute,
+                private confirm: ConfirmService,
                 private shelfService: ShelfService) {
         this.act.data.subscribe(value => {
             this.shelf = value.shelf;
@@ -22,8 +25,10 @@ export class ShelfDetailsComponent implements OnInit {
     }
 
     removeFromShelf(id: number) {
-        this.shelfService.removeBookFromShelf(id, this.shelf.id).subscribe(
-            () => this.getShelf()
+        this.confirm.confirm('Are you sure you want to remove this book from shelf?', () =>
+            this.shelfService.removeBookFromShelf(id, this.shelf.id).subscribe(
+                () => this.getShelf()
+            )
         )
     }
